rpi/requirements.txt
RPi.GPIO
picamera2
opencv-python-headless
numpy
